+++ docs/whats-new.rst
@@ -25,6 +25,11 @@ For convenience we also added a new property to the data fetcher that return the
     loader = ArgoDataFetcher().float(2901623)
     loader.domain  # Returns [89.093, 96.036, -0.278, 4.16, 15.0, 2026.0, numpy.datetime64('2010-05-14T03:35:00.000000000'),  numpy.datetime64('2013-01-01T01:45:00.000000000')]
 
+
+**Internals**
+
+- Update documentation pages to use the `xarray accessor sphinx extension <https://github.com/xarray-contrib/sphinx-autosummary-accessors>`_. (:pr:`104`) by `G. Maze <http://www.github.com/gmaze>`_.
+
 v0.1.8 (2 Nov. 2021)
 ---------------------
 
@@ -79,6 +84,8 @@ v0.1.8 (2 Nov. 2021)
     - ``trust_env`` to allow for local environment variables to be used by fsspec to connect to the internet. Usefull for those using a proxy.
 
 - Documentation on `Read The Docs` now uses a pip environment and get rid of memory eager conda. (:pr:`103`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+- :class:`xarray.Dataset` argopy accessor ``argo`` has a clean documentation.
 
 **Breaking changes with previous versions**